requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.1
django-crispy-forms==1.14.0
django-filter==21.1
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.10.1
Markdown==3.3.6
numpy==1.22.1
opencv-python==4.5.5.62
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0
zipp==3.7.0
